Custom Cursors & PickUp Alert
Changed the player's cursor to something more in tune with the game and modified it to change into a hand when hovering over an object you can pick up.
An alert with the item's name and image will spawn when you pick up an item now. It will not disappear atm but we can fix that later.